Automatically exit pipenv shell when leaving directory
If `~/test/Pipfile` exists and has a valid venv then starting at `~` and
cd'ing to `~/test` will start the pipenv shell as before.
Now when cd'ing to `~/test/foo` nothing will happen as before.

However when cd'ing to outside of `~/test`, eg. `~` or `~/test2` we will
automatically exit the pipenv shell and pass final directory back to the
original shell, which will then switch to that directory

This should fix #5.